third_party/WebKit/LayoutTests/screen_orientation/orientation-attribute.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/js-test.js"></script>
<script>
description("Basic test that accesses screen.orientation.");

shouldBeEqualToString("screen.orientation", "portrait-primary");
</script>
</body>
</html>
