third_party/WebKit/LayoutTests/screen_orientation/orientation-attribute-expected.txt

Basic test that accesses screen.orientation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS screen.orientation is "portrait-primary"
PASS successfullyParsed is true

TEST COMPLETE
